requirements.txt
streamlit
pillow
requests
python-dotenv
boto3
serpapi
google-search-results